Split Dashboard component
JIRA: RAIL-3383
