+++ libs/sdk-ui-dashboard/src/dashboard/index.ts
@@ -1,2 +1,3 @@
 // (C) 2021 GoodData Corporation
-export { IDashboardProps, Dashboard } from "./Dashboard";
+export { Dashboard } from "./Dashboard";
+export * from "./types";
